OCM-954 | Feat: small refactor to prepare for nodepool upgrades
