+++ cmd/dlt/machinepool/nodepool.go
@@ -11,9 +11,13 @@ import (
 func deleteNodePool(r *rosa.Runtime, nodePoolID string, clusterKey string, cluster *cmv1.Cluster) {
 	// Try to find the machine pool:
 	r.Reporter.Debugf("Loading machine pools for hosted cluster '%s'", clusterKey)
-	nodePool, err := r.OCMClient.GetNodePool(cluster.ID(), nodePoolID)
+	nodePool, exists, err := r.OCMClient.GetNodePool(cluster.ID(), nodePoolID)
 	if err != nil {
 		r.Reporter.Errorf("Failed to get machine pools for hosted cluster '%s': %v", clusterKey, err)
+		os.Exit(1)
+	}
+	if !exists {
+		r.Reporter.Errorf("Machine pool '%s' does not exist for hosted cluster '%s'", nodePoolID, clusterKey)
 		os.Exit(1)
 	}
