+++ pkg/ocm/upgrades.go
@@ -47,33 +47,6 @@ func (c *Client) GetUpgradePolicies(clusterID string) (upgradePolicies []*cmv1.U
 	return
 }
 
-func (c *Client) GetControlPlaneUpgradePolicies(clusterID string) (
-	controlPlaneUpgradePolicies []*cmv1.ControlPlaneUpgradePolicy,
-	err error) {
-	collection := c.ocm.ClustersMgmt().V1().
-		Clusters().
-		Cluster(clusterID).
-		ControlPlane().
-		UpgradePolicies()
-	page := 1
-	size := 100
-	for {
-		response, err := collection.List().
-			Page(page).
-			Size(size).
-			Send()
-		if err != nil {
-			return nil, handleErr(response.Error(), err)
-		}
-		controlPlaneUpgradePolicies = append(controlPlaneUpgradePolicies, response.Items().Slice()...)
-		if response.Size() < size {
-			break
-		}
-		page++
-	}
-	return
-}
-
 func (c *Client) GetScheduledUpgrade(clusterID string) (*cmv1.UpgradePolicy, *cmv1.UpgradePolicyState, error) {
 	upgradePolicies, err := c.GetUpgradePolicies(clusterID)
 	if err != nil {
@@ -98,36 +71,9 @@ func (c *Client) GetScheduledUpgrade(clusterID string) (*cmv1.UpgradePolicy, *cm
 	return nil, nil, nil
 }
 
-func (c *Client) GetControlPlaneScheduledUpgrade(clusterID string) (*cmv1.ControlPlaneUpgradePolicy, error) {
-	upgradePolicies, err := c.GetControlPlaneUpgradePolicies(clusterID)
-	if err != nil {
-		return nil, err
-	}
-	for _, upgradePolicy := range upgradePolicies {
-		if upgradePolicy.UpgradeType() == "ControlPlane" {
-			return upgradePolicy, nil
-		}
-	}
-
-	return nil, nil
-}
-
 func (c *Client) ScheduleUpgrade(clusterID string, upgradePolicy *cmv1.UpgradePolicy) error {
 	response, err := c.ocm.ClustersMgmt().V1().
 		Clusters().Cluster(clusterID).
-		UpgradePolicies().
-		Add().Body(upgradePolicy).
-		Send()
-	if err != nil {
-		return handleErr(response.Error(), err)
-	}
-	return nil
-}
-
-func (c *Client) ScheduleHypershiftControlPlaneUpgrade(clusterID string,
-	upgradePolicy *cmv1.ControlPlaneUpgradePolicy) error {
-	response, err := c.ocm.ClustersMgmt().V1().
-		Clusters().Cluster(clusterID).ControlPlane().
 		UpgradePolicies().
 		Add().Body(upgradePolicy).
 		Send()
@@ -147,16 +93,6 @@ func (c *Client) CancelUpgrade(clusterID string) (bool, error) {
 		UpgradePolicies().UpgradePolicy(scheduledUpgrade.ID()).
 		Delete().
 		Send()
-	if err != nil {
-		return false, handleErr(response.Error(), err)
-	}
-	return true, nil
-}
-
-func (c *Client) CancelControlPlaneUpgrade(clusterID, upgradeID string) (bool, error) {
-	response, err := c.ocm.ClustersMgmt().V1().
-		Clusters().Cluster(clusterID).ControlPlane().UpgradePolicies().
-		ControlPlaneUpgradePolicy(upgradeID).Delete().Send()
 	if err != nil {
 		return false, handleErr(response.Error(), err)
 	}
